// src/typings/api/c2c/advertise/post-advertise.d.ts
export interface IMainChainAddr {
  id:string;
  address?: any;
  name: string;
  memo?: any;
}

interface IPaymentDetails {
    businessId: string;
    uid: string;
    /** 法币 ID */
    legalCurrencyId: string;
    /** 收付类型 */
    paymentTypeCd: string;
    /** 姓名 */
    name: string;
    /** 账号 */
    account: string;
    /** 开户行 */
    bankOfDeposit: string;
    /** 二维码 (图片资源地址) */
    qrCodeAddr: string;
    /** 是否启用，1，启用，0，未启用 */
    enabled:number
    /** 是否显示灰色 */
    grey: boolean;
    /** 付款详情 */
    paymentDetails?: string
    /** 银行支行 */
    bankBranch?: string
    /** 唯一标识 */
    id: string;
}

export interface IAdvertList {
    /** 最小金额 */
    minAmount: string;
    /** 有效期 */
    validDate: number;
    /** 支付方式 */
    payments: string[];
    /** 备注 */
    remark: string;
    /** 服务人数，广告主才有 */
    customerCnt: number;
    /** 成单数大于 */
    completedOrderCount: number;
    /** 广告方向 (BUY SELL) */
    advertDirectCd: string;
    /** 下架时间 */
    removalTime: number;
    /** 价格 */
    price: string;
    /** 最大金额 */
    maxAmount: string;
    /** 下架详情 */
    removalReason: string;
    /** 数量 */
    quantity: string;
    /** 昵称 */
    nickName?: any;
    /** 订单数 */
    orderCount: number;
    /** 是否是商户 */
    merchant: boolean;
    /** 认证等级 */
    certificationLevelCd: string;
    /** 头像 */
    avatar?: any;
    /** 广告 id */
    advertId: string;
    /** 是否在线 */
    onLine: boolean;
    /** 币种 id */
    coinId: string;
    /** 交易总量 */
    totalOrderAmount: string;
    /** 创建时间 */
    createTime: number;
    /** 主链地址 */
    mainchainAddrs: IMainChainAddr[];
    /** 币种名称 */
    coinName: string;
    /** 交易类型 */
    tradeTypeCd: string;
    /** 支付方式详情 */
    paymentDetails?: IPaymentDetails[];
    /** 广告状态（ON_SHELVES 上架中、EXPIRED 到期下架、TAKE_OFF_SHELVES_MERCHANT 商家下架、TAKE_OFF_SHELVES_SYSTEM 系统下架、DEALING 交易中、WIDTHDRAW 撤回中） */
    statusCd: string;
    /** 区域名字 */
    areaName: string;
    /** 法币符号 */
    currencySymbol: string;
    /** 新的广告状态（跟商户关联）(RESTING 休息中、NORMAL 正常、DEALING 交易中) */
    advertNewStatus: string;
    /** 是否置灰（重新上架） */
    grey: boolean;
    /** 是否可交易 */
    canTrade: boolean;
    /** 新增支付方式（返回是否需要置灰）*/
    newPayments: INewPayments[]
}
export interface INewPayments {
    /** BANK、WECHAT、ALIPAY */
    name: string
    /** 是否需要显示灰色 */
    grey: boolean
}